modified the readme file and implemented the KNN classifier and regressor models.
